requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
confuse==2.0.1
curl_cffi==0.13.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
feedparser==6.0.12
frozendict==2.4.6
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
multitasking==0.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.4
openai==2.6.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.2
pefile==2023.2.7
pickleshare==0.7.5
platformdirs==4.5.0
playwright==1.48.0
playwright-stealth==2.0.0
prompt_toolkit==3.0.52
protobuf==6.33.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==12.0.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-amazon-sp-api==1.9.50
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yfinance==0.2.66
